# domain/employee.go
package domain

import "time"

type Login struct {
	Phone    string `json:"phone"`
	Password string `json:"password"`
}

type LoginResponse struct {
	Token       string                 `json:"token"`
	Employee    Employee               `json:"employee"`
	Permissions map[string]interface{} `json:"permissions"`
}

type Employee struct {
	Id        string     `gorm:"id" json:"id"`
	RoleId    string     `gorm:"role_id" json:"role_id"`
	FirstName string     `gorm:"first_name" json:"first_name"`
	LastName  string     `gorm:"last_name" json:"last_name"`
	Email     string     `gorm:"email" json:"email"`
	Phone     string     `gorm:"phone" json:"phone"`
	Password  string     `gorm:"password" json:"password"`
	Language  string     `gorm:"language" json:"language"`
	CreatedAt *time.Time `gorm:"created_at" json:"created_at"`
	UpdatedAt *time.Time `gorm:"updated_at" json:"updated_at"`
}

type EmployeeRequest struct {
	Id        string `gorm:"id" json:"-"`
	RoleId    string `gorm:"role_id" json:"role_id"`
	FirstName string `gorm:"first_name" json:"first_name"`
	LastName  string `gorm:"last_name" json:"last_name"`
	Email     string `gorm:"email" json:"email"`
	Phone     string `gorm:"phone" json:"phone"`
	Password  string `gorm:"password" json:"password"`
	Language  string `gorm:"language" json:"language"`
}